financial_analyst_agent/backend/requirements.txt
crewai
crewai-tools
python-dotenv
fastapi
pydantic
langchain-community